Data Science intensive course. Capstone Project
This project deals with the always painful process of moving to a new
city. A recommender model based on an unsupervised approach is here
presented for the city of Washington DC (US).
The work includes the data wrangling process, a statistical analysis of
the criminal activity recorded along 2011 to 2013 and finally the use
of machine learning to build an unsupervised recommender system based
on the preferences of the user and on four main features: safety,
access to public transportation services, education and economic
indexes.